Add BGPView Cat tool (`bvcat`)
A tool to dump binary views (as created by the archiver) to stdout (a la `cat`). Unfortunately it is pretty inefficient as it needs to read the entire view into memory before dumping it, but to dump at the same time as reading would require a bunch more work that I'm not willing to do at the moment. However, since the view is re-used, it should be faster to read views 1..n (i.e., reading the first view will be the slowest).