Add "bsrt" (bgpstream routingtable) io module
Copy large parts of bgpcorsaro framework and its routingtables plugin
into the new "bsrt" io module, trying to keep changes to a minimum to
make it work with bgpview.  Mainly, this involved turning the bgpcorsaro
logic inside out so the bgpview generated at end-of-interval can instead
be fetched by recv_view().  libroutingtables is unchanged.  A lot of
cleanup and optimization remains to be done.