extend moas_signature struct for dynamic origins
the `moas_signature` struct changes from containing a static allocated
int32 array of size 128 and a count integer, to a static array of size 4
and a dynamic array and a origins count. the dynamic array is allocated
if there are more than 4 origins in the moas event and is assigned with
a static 128 size.

this commit is designed to reduce the memory consumption of the moas
consumer by reducing the sizes of the moas signatures needed to be
stored in memory.