per-geo-visibility: fix bad mask lens for v6 prefixes > 64 bits
Mask lengths were being incorrectly calculated to be zero, due
to undefined behaviour when evaluating log2(0).

This should also fix the assertion failure that has been
crashing the consumer recently in production:

    bgpstream_utils_patricia.c:677: bgpstream_patricia_tree_insert: Assertion `bgpstream_pfx_equal(&node_it->prefix, pfx)' failed.

which was caused by a failed attempt to insert one of the
erroneous /0 prefixes into a patricia trie.