pfx-origins: allow origin arrays to grow dynamically
64 entries is more than enough for most prefixes, but I'm seeing
the occasional prefix with over 100 possible origins.

Rather than just bumping the array size to 128 (and therefore
doubling the memory footprint), this change will only increase
the array size for prefixes that need it, in increments of 64
array entries at a time.

Still undergoing long-term testing for memory leaks, but initial
signs are good so far...