Edges consumer: treat edges as directional
Previously, edges A->B and B->A were treated as the same
edge but this means that we lose information about the
directionality of the observed edges.

Now, A->B and B->A are treated as distinct edges and "NEW"
records will created accordingly when each instance appears
for the first time.